refactor(scm): consolidate dual GitHub clients — delete dogfooding/github-client (#2553)
scm/index.ts has documented since #1136 that it replaces the prior
dual-path GitHub clients. Half of that migration shipped (the
self-development client was deleted in #2402); the dogfooding shim
never went away. Finishing it now.

What stayed of dogfooding/github-client.ts was just two pure URL
parsers (parsePRUrl, parseIssueUrl) — the full GitHubClient class,
GitHubError, createGitHubClientFromEnv, and GitHubClientConfig had
zero non-test consumers. Lifted the URL parsers into scm/, migrated
the two real callers, deleted the rest.

Changes:
- New scm/url-parsers.ts + scm/url-parsers.test.ts (URL parsing logic
  lifted verbatim from github-client.ts; tests rewritten as a focused
  suite covering the same valid/invalid cases).
- scm/index.ts re-exports parsePRUrl + parseIssueUrl.
- dogfooding/pr-reviewer.ts: imports parsePRUrl from scm/url-parsers.
- dogfooding/issue-triage.ts: imports parseIssueUrl from scm/url-parsers.
- dogfooding/pr-reviewer.test.ts: same import update.
- dogfooding/index.ts: drops the entire github-client re-export block
  (GitHubClient, GitHubError, createGitHubClientFromEnv,
  GitHubClientConfig). Zero external consumers, confirmed by grep.
- DELETE dogfooding/github-client.ts + dogfooding/github-client.test.ts.
- scm/github-provider.ts + scm/index.ts: rewrite the "replaces previous
  dual-path GitHub clients" docstring now that both replacements have
  shipped.

Net: −1300+ lines (most of github-client.ts + its test suite).

Verified:
- pnpm typecheck clean
- pnpm vitest run src/scm/ src/dogfooding/ → 187/187 pass

Closes #2553.
